Rename pad alignment commands
