+++ src/main/java/com/stuypulse/robot/commands/drivetrain/DrivetrainPadAlignV2.java
@@ -6,11 +6,15 @@
 package com.stuypulse.robot.commands.drivetrain;
 
 import com.stuypulse.stuylib.control.Controller;
+import com.stuypulse.stuylib.math.Angle;
 import com.stuypulse.stuylib.streams.IFuser;
 import com.stuypulse.stuylib.streams.booleans.BStream;
 import com.stuypulse.stuylib.streams.booleans.filters.BDebounceRC;
+import com.stuypulse.stuylib.streams.filters.IFilter;
+import com.stuypulse.stuylib.streams.filters.LowPassFilter;
 
 import com.stuypulse.robot.commands.ThenShoot;
+import com.stuypulse.robot.constants.Settings;
 import com.stuypulse.robot.constants.Settings.Alignment;
 import com.stuypulse.robot.constants.Settings.Limelight;
 import com.stuypulse.robot.subsystems.Camera;
@@ -20,32 +24,52 @@ import com.stuypulse.robot.subsystems.Drivetrain;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class OldPadAlign extends CommandBase {
+public class DrivetrainPadAlignV2 extends CommandBase {
 
     private final Drivetrain drivetrain;
 
     private final BStream finished;
 
-    private final IFuser angleError;
-    protected final Controller angleController;
+    private IFilter speedAdjFilter;
 
-    public OldPadAlign(Drivetrain drivetrain, Camera camera) {
+    private final IFuser angleError;
+    private final IFuser distanceError;
+
+    protected final Controller angleController;
+    protected final Controller distanceController;
+
+    public DrivetrainPadAlignV2(Drivetrain drivetrain, Camera camera) {
         this.drivetrain = drivetrain;
 
         // find errors
         angleError =
                 new IFuser(
                         Alignment.FUSION_FILTER,
-                        () -> camera.getXAngle().toDegrees(),
+                        () -> camera.getXAngle()
+                                .add(Angle.fromDegrees(Limelight.PAD_YAW.get()))
+                                .toDegrees(),
                         () -> drivetrain.getRawGyroAngle());
 
+        distanceError =
+                new IFuser(
+                        Alignment.FUSION_FILTER,
+                        () -> Settings.Limelight.PAD_SHOT_DISTANCE - camera.getDistance(),
+                        () -> drivetrain.getDistance());
+
         // handle errors
+        speedAdjFilter = new LowPassFilter(Alignment.SPEED_ADJ_FILTER);
         this.angleController = Alignment.Angle.getController();
+        this.distanceController = Alignment.Speed.getController();
 
         // finish optimally
         finished =
                 BStream.create(camera::hasTarget)
+                        .and(
+                                () ->
+                                        Math.abs(drivetrain.getVelocity())
+                                                < Limelight.MAX_VELOCITY.get())
                         .and(() -> angleController.isDone(Limelight.MAX_ANGLE_ERROR.get()))
+                        .and(() -> distanceController.isDone(Limelight.MAX_DISTANCE_ERROR.get()))
                         .filtered(new BDebounceRC.Rising(Limelight.DEBOUNCE_TIME));
 
         addRequirements(drivetrain);
@@ -55,16 +79,29 @@ public class OldPadAlign extends CommandBase {
     public void initialize() {
         drivetrain.setLowGear();
 
+        speedAdjFilter = new LowPassFilter(Alignment.SPEED_ADJ_FILTER);
+
         angleError.initialize();
+        distanceError.initialize();
+    }
+
+    private double getSpeedAdjustment() {
+        double error = angleError.get() / Limelight.MAX_ANGLE_FOR_MOVEMENT.get();
+        return speedAdjFilter.get(Math.exp(-error * error));
+    }
+
+    private double getSpeed() {
+        double speed = distanceController.update(distanceError.get());
+        return speed * getSpeedAdjustment();
     }
 
     private double getTurn() {
-        return angleController.update(angleError.get() - 0.5 * Limelight.LIMELIGHT_YAW.get());
+        return angleController.update(angleError.get());
     }
 
     @Override
     public void execute() {
-        drivetrain.arcadeDrive(0, getTurn());
+        drivetrain.arcadeDrive(getSpeed(), getTurn());
     }
 
     @Override
